+++ package.json
@@ -15,6 +15,8 @@
     "@types/node": "^20.3.1",
     "@types/react": "^18.2.12",
     "@types/react-dom": "^18.2.5",
+    "@upstash/redis": "^1.21.0",
+    "@vercel/edge": "^0.3.4",
     "@zolplay/react": "^0.5.1",
     "@zolplay/utils": "^1.3.4",
     "autoprefixer": "^10.4.14",
