--- src/app/api/cdg1/route.ts
@@ -1,7 +1,0 @@
-import { createHandler } from '~/helpers/factory'
-
-export const fetchCache = 'force-no-store'
-export const runtime = 'edge'
-export const preferredRegion = 'cdg1'
-
-export const POST = createHandler('cdg1')
