--- src/app/api/cle1/route.ts
@@ -1,7 +1,0 @@
-import { createHandler } from '~/helpers/factory'
-
-export const fetchCache = 'force-no-store'
-export const runtime = 'edge'
-export const preferredRegion = 'cle1'
-
-export const POST = createHandler('cle1')
