--- src/app/api/icn1/route.ts
@@ -1,7 +1,0 @@
-import { createHandler } from '~/helpers/factory'
-
-export const fetchCache = 'force-no-store'
-export const runtime = 'edge'
-export const preferredRegion = 'icn1'
-
-export const POST = createHandler('icn1')
