+++ src/app/api/regions/cdg1/route.ts
@@ -1,0 +1,9 @@
+
+import { createHandler } from '~/helpers/factory'
+
+export const fetchCache = 'force-no-store'
+export const runtime = 'edge'
+export const preferredRegion = 'cdg1'
+
+export const POST = createHandler('cdg1')
+
