+++ src/app/api/regions/dub1/route.ts
@@ -1,0 +1,9 @@
+
+import { createHandler } from '~/helpers/factory'
+
+export const fetchCache = 'force-no-store'
+export const runtime = 'edge'
+export const preferredRegion = 'dub1'
+
+export const POST = createHandler('dub1')
+
