+++ src/app/api/regions/hkg1/route.ts
@@ -1,0 +1,9 @@
+
+import { createHandler } from '~/helpers/factory'
+
+export const fetchCache = 'force-no-store'
+export const runtime = 'edge'
+export const preferredRegion = 'hkg1'
+
+export const POST = createHandler('hkg1')
+
