+++ src/app/api/regions/hnd1/route.ts
@@ -1,0 +1,9 @@
+
+import { createHandler } from '~/helpers/factory'
+
+export const fetchCache = 'force-no-store'
+export const runtime = 'edge'
+export const preferredRegion = 'hnd1'
+
+export const POST = createHandler('hnd1')
+
