+++ src/app/api/regions/lhr1/route.ts
@@ -1,0 +1,9 @@
+
+import { createHandler } from '~/helpers/factory'
+
+export const fetchCache = 'force-no-store'
+export const runtime = 'edge'
+export const preferredRegion = 'lhr1'
+
+export const POST = createHandler('lhr1')
+
