+++ src/app/api/token/GET.ts
@@ -1,0 +1,10 @@
+import { ipAddress } from '@vercel/edge'
+import { NextResponse, type NextRequest } from 'next/server'
+
+export async function GET(req: NextRequest) {
+  const ip = ipAddress(req) || 'unknown'
+
+  return NextResponse.json({
+    ip,
+  })
+}
