+++ src/app/api/token/route.ts
@@ -1,0 +1,3 @@
+export { GET } from './GET'
+
+export const runtime = 'edge'
